Fix CUDA

+++ src/beam_search_scorer.cpp
@@ -195,7 +195,7 @@ void BeamSearchScorer::Finalize(Sequences& sequences,
   // sequence_scores is the optional Score of each sequence, with shape (batch_size * num_return_sequences).
 
   // Finalize all open beam hypotheses and add to generated hypotheses.
-  for (int batch_index = 0; batch_index < batch_size_; batch_index++) {
+  for (size_t batch_index = 0; batch_index < batch_size_; batch_index++) {
     BeamHypotheses& beam_hyp = beam_hyps_[batch_index];
     if (beam_hyp.done_) {
       continue;
